requirements.txt
-i https://pypi.org/simple
django-rest-framework==0.1.0
django==2.2.7
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
